requirements.txt
PyPDF2
sentence-transformers
python-dotenv
azure-search-documents
streamlit
openai